requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.0
pandas==2.2.2
pillow==10.4.0
PyJWT==2.9.0
PyKakao==0.0.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyupbit==0.2.34
requests==2.32.3
schedule==1.2.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
websockets==13.0.1
Werkzeug==3.0.4
